In rootless + cgroupsv1, resource limits are an error
However, in some cases (unset limits), we can completely remove
the limit and avoid errors. This works around a bug where the
Podman frontend is setting a Pids limit of 0 on some rootless
systems.

For now, this is only implemented for the PID limit. It can
easily be extended to other resource limits, but it is a fair bit
of code to do so, so I leave that exercise to someone else.

Fixes #6834

Signed-off-by: Matthew Heon <matthew.heon@pm.me>