[CI:IMG] Bump crun to 1.14 on CI VMs
This began as an empty commit intended solely to get CI to
rebuild the VMs using crun 1.14 (up from 1.13). Twin goals:

  1) Be able to test #6693 (--sdnotify option); and
  2) Get rid of 'cgroup.freeze' CI flakes.

These are fixed by crun PRs 419 and 423 respectively.

CI failed on the original PR submission, with errors on ginkgo
install. The change to Makefile is intended to address that.

The change to setup_environment.sh is intended to address
a flake we're frequently seeing with the Fedora dnf repos.
This adds a retry in case of a failing dnf command.

Signed-off-by: Ed Santiago <santiago@redhat.com>