make binaries.coverage
Add make targets to build podman binaries (local & remote) that are
instrumented to generate coverage data.  The binaries are placed in
`./bin` and are suffixed with `.coverage` to distinguish them from
ordinary builds.

Both binaries can later be used to collect coverage profiles, which can
further be sent over to codecov.io to assemble a global coverage
profile.

NOTE: there are certain limitations when collecting coverage data,
mainly that it it's only possible if the binary executes gracefully and
without `os.Exit(...)`.

Signed-off-by: Valentin Rothberg <rothberg@redhat.com>