Lower missing runtime from Errorf to Infof
Creating a container with --runtime=...  Can cause a situation where all of the tools that look at containers starts generating an ugly error message that
does not effect the running of the tool, but gives the users a bad experience.

Also users can not fix this situation easily without removing the container or
figuring out how to edit containers.conf.

From looking at the problem, it does not look like a real issue and should just be printed at the Info level.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>